Split into files main.cpp and transformed main
