+++ skuratov.danil/T2/delimiter.cpp
@@ -1,0 +1,17 @@
+#include "delimiter.hpp"
+
+std::istream& skuratov::operator>>(std::istream& in, DelimiterI&& exp)
+{
+  std::istream::sentry guard(in);
+  if (!guard)
+  {
+    return in;
+  }
+  char c = 0;
+  in >> c;
+  if (c != exp.expected)
+  {
+    in.setstate(std::ios::failbit);
+  }
+  return in;
+}
