+++ skuratov.danil/T2/delimiter.hpp
@@ -1,0 +1,15 @@
+#ifndef DELIMITER_HPP
+#define DELIMITER_HPP
+
+#include <iostream>
+
+namespace skuratov
+{
+	struct DelimiterI
+	{
+		char expected;
+	};
+	std::istream& operator>>(std::istream& in, DelimiterI&& exp);
+}
+
+#endif
