+++ skuratov.danil/T2/main.cpp
@@ -1,99 +1,26 @@
-#include <iostream>
 #include <limits>
+#include <list>
+#include <iterator>
+#include <algorithm>
 
-namespace skuratov
-{
-  struct Data
-  {
-    Data(int aa, int bb):
-      a(aa),
-      b(bb)
-    {}
-    int getA() const
-    {
-      return a;
-    }
-    int getB() const
-    {
-      return b;
-    }
-  private:
-    int a, b;
-  };
-
-  struct DelimiterI
-  {
-    char expected;
-  };
-
-  std::istream& operator>>(std::istream& in, DelimiterI&& exp)
-  {
-    std::istream::sentry guard(in);
-    if (!guard)
-    {
-      return in;
-    }
-    char c = 0;
-    in >> c;
-    if (c != exp.expected)
-    {
-      in.setstate(std::ios::failbit);
-    }
-    return in;
-  }
-
-  std::istream& operator>>(std::istream& in, Data& value)
-  {
-    std::istream::sentry guard(in);
-    if (!guard)
-    {
-      return in;
-    }
-
-    using del = DelimiterI;
-    int a = 0, b = 0;
-    in >> del{'['} >> a >> del{';'} >> b >> del{']'};
-
-    if (in)
-    {
-      value = Data(a, b);
-    }
-    return in;
-  }
-  std::ostream& operator<<(std::ostream& out, const Data& value)
-  {
-    std::ostream::sentry guard(out);
-    if (!guard)
-    {
-      return out;
-    }
-    out << '[' << value.getA() << ";" << value.getB() << ']';
-    return out;
-  }
-}
+#include "delimiter.hpp"
+#include "dataStruct.hpp"
 
 int main()
 {
-  skuratov::Data d(0, 0);
-  if (!(std::cin >> d))
+  using namespace skuratov;
+  using inputItT = std::istream_iterator< DataStruct >;
+  std::list < DataStruct > d(inputItT{ std::cin }, inputItT{});
+  while (!(std::cin.eof()))
   {
-    std::cin.clear();
-    std::cin.ignore(std::numeric_limits< std::streamsize >::max(), '\n');
-    if (!(std::cin >> d))
+    std::copy(inputItT{ std::cin }, inputItT{}, std::back_inserter(d));
+    if (std::cin.fail())
     {
-      std::cerr << "Error\n";
-      return 1;
+      std::cin.clear();
+      std::cin.ignore(std::numeric_limits< std::streamsize >::max(), '\n');
     }
   }
-  std::cout << d << "\n";
-  return 0;
-
-  /*
-  //yamain.cpp
-  using input_it_t = std::istream_iterator< DataStruct >;
-  std::vector< DataStruct > data(input_it_t{std::cin}, input_it_t{});
-
-  using output_it_t = std::ostream_iterator< DataStruct >;
-  std::copy(data.cbegin(), data.cend(), output_it_t{std::cout, "\n"});
-  */
+  d.sort();
+  using outputItT = std::ostream_iterator< DataStruct >;
+  std::copy(d.cbegin(), d.cend(), outputItT{ std::cout, "\n" });
 }
